chore: move x86 tests into separate file
